Move dom to top-level.
